Corrected CSS issue to allow footer to stick to bottom
The header changes sizes when switching to the mobile view and on the projects.html page there is not enough content for the <main> element to extend beyond the vertical height of the page. This issue was corrected by setting a minimum height for the header so that it would not become smaller.